add t3/t4

+++ InstaMedData/Models/Tests/T3T4.cs
@@ -7,8 +7,8 @@ namespace InstaMedData.Models
     public class T3T4
     {
         public int Id { get; set; }
-        public float T3Result { get; set; }
-        public float T4Result { get; set; }
+        public decimal T3Result { get; set; }
+        public decimal T4Result { get; set; }
         public String DocName { get; set; }
         public bool isT3T4 { get; set; }
